requirements.txt
streamlit>=1.30.0
tensorflow-cpu==2.12.0
pillow
pandas
numpy
scikit-learn
opencv-python
requests
